moved book and page out from under key.  fixes some wording and fixed jsonrpc query
